Split off OptimizeProjections
